Change GPIO semantics and fully rewrite I2C
Before, an I2C bus was queueing all reads and writes in an agent. As it
is a "general purpose" library, this has been abandonned.  The new fns
are much simpler and straightforward but aren't thread-safe anymore.
Concurrency, if needed, must be taken care of by the user who will pick
the right tool (agents, core.async, manifolds, etc).